+++ src/test/clojure/clojure/tools/analyzer/test/analyzer.clj
@@ -26,7 +26,7 @@
                     require))
            :exprs last :var)
          #'clojure.core/require)) ;should be `clojure.set/intersection`
-  (is (analyze-ns (pb-reader-for-ns 'clojure.tools.analyzer.test.require) 'my-ns 'my-ns)))
+  (is (analyze-ns 'clojure.tools.analyzer.test.require)))
 
 (deftest records-test
   (is (analyze-ns 'clojure.tools.analyzer.test.records)))