+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+0.5.0 - Released 22 September 2013
+BREAKING CHANGES
+- Change interface to analyze-ns
+  - takes a namespace symbol and options. See docstring
+
+ENHANCEMENTS
+- add hinted fields to deftype*
+- add analyze-file
+
 0.4.3
 - Support Clojurescript